New Motor controller architecture with feedforward

+++ 1_Code/3_STM32/0_workspace/X1_Thorn_v3/Core/Src/User/Peripheral/ESC/Motor_PI.h
@@ -31,7 +31,7 @@ typedef struct {
  * @param  integratorMin/Max  Limits on integrator term (anti‐windup)
  * @param  tau       Derivative filter time constant (if zero, do plain D)
  */
-void Motor_PID_Init(pid_handle_t *pid, float Kp, float Ki, float dt, uint16_t outputMin, uint16_t outputMax);
+void Motor_PID_Init(pid_handle_t *pid, float Kp, float Ki, float dt, float outputMin, float outputMax);
 
 /**
  * @brief  Compute PID output given setpoint and measurement
@@ -40,6 +40,6 @@ void Motor_PID_Init(pid_handle_t *pid, float Kp, float Ki, float dt, uint16_t ou
  * @param  measurement Current measured value
  * @retval control output (clamped between outputMin, outputMax)
  */
-uint16_t Motor_PID_Update(pid_handle_t *pid, float measurement, float setpoint);
+float Motor_PID_Update(pid_handle_t *pid, float measurement, float setpoint);
 
 #endif /* SRC_USER_PERIPHERAL_ESC_MOTOR_PI_H_ */
